refactor: objects와 helpers로 클래스 특징에 맞게 분리
