+++ src/objects/LottoCompany.js
@@ -1,5 +1,5 @@
-import InputManager from './InputManager.js';
-import { generateMapWithZeroValue } from './lib/utils.js';
+import { InputManager } from '../helpers/index.js';
+import { generateMapWithZeroValue } from '../lib/utils.js';
 
 class LottoCompany {
   static #RANKS = [1, 2, 3, 4, 5];
