+++ src/objects/index.js
@@ -1,0 +1,4 @@
+export { default as Lotto } from './Lotto.js';
+export { default as LottoBuyer } from './LottoBuyer.js';
+export { default as LottoCompany } from './LottoCompany.js';
+export { default as LottoShop } from './LottoShop.js';
